+++ clinica-odontologica/clinica-odontologica/src/main/java/com/backend/clinicaodontologica/dto/entrada/turno/TurnoEntradaDto.java
@@ -11,10 +11,10 @@ import java.time.LocalDateTime;
 public class TurnoEntradaDto {
 
     @NotNull(message = "El paciente no puede ser nulo")
-    private int pacienteId;
+    private Long pacienteId;
 
     @NotNull(message = "El odontologo no puede ser nulo")
-    private int odontologoId;
+    private Long odontologoId;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
     @FutureOrPresent(message = "La fecha no puede ser anterior al día de hoy")
@@ -24,25 +24,25 @@ public class TurnoEntradaDto {
     public TurnoEntradaDto() {
     }
 
-    public TurnoEntradaDto(int pacienteId, int odontologoId, LocalDateTime fechaYHora) {
+    public TurnoEntradaDto(Long pacienteId, Long odontologoId, LocalDateTime fechaYHora) {
         this.pacienteId = pacienteId;
         this.odontologoId = odontologoId;
         this.fechaYHora = fechaYHora;
     }
 
-    public int getPacienteId() {
+    public Long getPacienteId() {
         return pacienteId;
     }
 
-    public void setPacienteId(int pacienteId) {
+    public void setPacienteId(Long pacienteId) {
         this.pacienteId = pacienteId;
     }
 
-    public int getOdontologoId() {
+    public Long getOdontologoId() {
         return odontologoId;
     }
 
-    public void setOdontologoId(int odontologoId) {
+    public void setOdontologoId(Long odontologoId) {
         this.odontologoId = odontologoId;
     }
