+++ clinica-odontologica/clinica-odontologica/src/main/java/com/backend/clinicaodontologica/service/IPacienteService.java
@@ -12,9 +12,9 @@ public interface IPacienteService {
 
     PacienteSalidaDto registrarPaciente(PacienteEntradaDto paciente);
 
-    PacienteSalidaDto buscarPacientePorId(int id);
+    PacienteSalidaDto buscarPacientePorId(Long id);
 
-    void eliminarPaciente(int id);
+    void eliminarPaciente(Long id);
 
     PacienteSalidaDto modificarPaciente(PacienteModificacionEntradaDto pacienteModificado);
