+++ clinica-odontologica/clinica-odontologica/src/main/java/com/backend/clinicaodontologica/service/ITurnoService.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface ITurnoService {
     TurnoSalidaDto registrarTurno(TurnoEntradaDto turnoEntradaDto);
     List<TurnoSalidaDto> listarTurnos();
-    TurnoSalidaDto buscarTurnoPorId(int id);
-    void eliminarTurno(int id);
+    TurnoSalidaDto buscarTurnoPorId(Long id);
+    void eliminarTurno(Long id);
     //TurnoSalidaDto modificarTurno(completaraquieldto);
 
 }
